chore: rename dashboard.domain
