+++ frontend/core/containers/thread/DashboardThread/logic/useBaseInfo/index.ts
@@ -4,7 +4,7 @@ import useCommunity from '~/hooks/useCommunity'
 import useDashboard from '~/hooks/useDashboard'
 import useQuery from '~/hooks/useQuery'
 import type { TCommunity, TDsbBaseInfoRoute, TEditFunc } from '~/spec'
-import type { TDsbFields } from '~/stores/dashboard.domain/spec'
+import type { TDsbFields } from '~/stores/dashboard/spec'
 import { BASEINFO_KEYS } from '../../constant'
 import S from '../../schema'
 import useHelper from '../useHelper'
