+++ frontend/core/containers/thread/DashboardThread/logic/useBroadcast.ts
@@ -1,7 +1,7 @@
 import { pick } from 'ramda'
 import useDashboard from '~/hooks/useDashboard'
 import type { TBroadcastConf, TBroadcastLayout, TDsbBroadcastRoute, TEditFunc } from '~/spec'
-import { FIELD } from '~/stores/dashboard.domain/constant'
+import { FIELD } from '~/stores/dashboard/constant'
 
 import useHelper from './useHelper'
