+++ frontend/core/containers/thread/DashboardThread/logic/useHelper/useTouch.ts
@@ -1,7 +1,7 @@
 import { any, equals } from 'ramda'
 import useDashboard from '~/hooks/useDashboard'
 
-import type { TDsbField } from '~/stores/dashboard.domain/spec'
+import type { TDsbField } from '~/stores/dashboard/spec'
 
 export type TRet = {
   isChanged: (field: TDsbField) => boolean
