+++ frontend/core/stores/provider.landing.tsx
@@ -3,7 +3,7 @@ import METRIC from '~/const/metric'
 import type { TCommunity } from '~/spec'
 import AccountStoreProvider from '~/stores/account.domain/provider'
 import CommunityStoreProvider from '~/stores/community/provider'
-import DashboardStoreProvider from '~/stores/dashboard.domain/provider'
+import DashboardStoreProvider from '~/stores/dashboard/provider'
 import GeneralStoreProvider from '~/stores/general/provider'
 import ThemeStoreProvider from '~/stores/theme.domain/provider'
 import WallpaperStoreProvider from '~/stores/wallpaper.domain/provider'
