+++ frontend/core/stores/provider.tsx
@@ -2,7 +2,7 @@ import type { FC } from 'react'
 import type { TCommunity } from '~/spec'
 import AccountStoreProvider from '~/stores/account.domain/provider'
 import CommunityStoreProvider from '~/stores/community/provider'
-import DashboardStoreProvider from '~/stores/dashboard.domain/provider'
+import DashboardStoreProvider from '~/stores/dashboard/provider'
 import GeneralStoreProvider from '~/stores/general/provider'
 import ThemeStoreProvider from '~/stores/theme.domain/provider'
 import WallpaperStoreProvider from '~/stores/wallpaper.domain/provider'
